+++ KodlaKariyerineBaslaJavaBootcamp/src/Hafta4/java-bootcamp-week4/src/main/java/com/example/javabootcampweek4/service/PlaneService.java
@@ -7,8 +7,6 @@ import com.example.javabootcampweek4.requestDTO.PlaneSaveRequestDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDate;
-import java.util.List;
 
 @Service
 public class PlaneService {
@@ -19,18 +17,18 @@ public class PlaneService {
     @Autowired
     PlaneConverter planeConverter;
 
-    public Long saveTHY(PlaneSaveRequestDTO planeSaveRequestDTO) {
+    public Integer saveTHY(PlaneSaveRequestDTO planeSaveRequestDTO) {
         THY thyPlane = planeConverter.convertFromPlaneSaveRequestDTOToTHY(planeSaveRequestDTO);
         thyPlane = planeRepository.save(thyPlane);
         return thyPlane.getPlaneId();
 
     }
-    public Long savePegasus(PlaneSaveRequestDTO planeSaveRequestDTO){
+    public Integer savePegasus(PlaneSaveRequestDTO planeSaveRequestDTO){
         Pegasus pegasusPlane = planeConverter.convertFromPlaneSaveRequestDTOToPegasus(planeSaveRequestDTO);
         pegasusPlane = planeRepository.save(pegasusPlane);
         return pegasusPlane.getPlaneId();
     }
-    public Long saveOnurAir(PlaneSaveRequestDTO planeSaveRequestDTO){
+    public Integer saveOnurAir(PlaneSaveRequestDTO planeSaveRequestDTO){
         OnurAir onurAirPlane = planeConverter.convertFromPlaneSaveRequestDTOToOnurAir(planeSaveRequestDTO);
         onurAirPlane = planeRepository.save(onurAirPlane);
         return onurAirPlane.getPlaneId();
